backend/requirements.txt
django
djangorestframework
django-filter
sqlparse
wheel